ScImgDataLoader_PSD: use 64bit file offsets
git-svn-id: svn://scribus.net/branches/Version15x/Scribus@25355 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/imagedataloaders/scimgdataloader_psd.h
@@ -47,13 +47,13 @@ protected:
 	bool IsValid( const PSDHeader & header ) const;
 	bool IsSupported( const PSDHeader & header ) const;
 
-	bool LoadPSD( QDataStream & s, const PSDHeader & header);
-	bool LoadPSDResources( QDataStream & s, const PSDHeader & header, uint dataOffset );
-	bool LoadPSDImgData( QDataStream & s, const PSDHeader & header, uint dataOffset );
-	bool loadChannel( QDataStream & s, const PSDHeader & header, QList<PSDLayer> &layerInfo, uint layer, int channel, int component, RawImage &tmpImg);
-	bool loadLayerChannels( QDataStream & s, const PSDHeader & header, QList<PSDLayer> &layerInfo, uint layer, bool* firstLayer);
-	bool loadLayer( QDataStream & s, const PSDHeader & header);
-	bool parseLayer( QDataStream & s, const PSDHeader & header);
+	bool LoadPSD(QDataStream& s, const PSDHeader& header);
+	bool LoadPSDResources(QDataStream& s, const PSDHeader& header, qint64 dataOffset);
+	bool LoadPSDImgData(QDataStream& s, const PSDHeader& header, qint64 dataOffset);
+	bool loadChannel(QDataStream& s, const PSDHeader& header, QList<PSDLayer>& layerInfo, uint layer, int channel, int component, RawImage& tmpImg);
+	bool loadLayerChannels(QDataStream& s, const PSDHeader& header, QList<PSDLayer>& layerInfo, uint layer, bool* firstLayer);
+	bool loadLayer(QDataStream& s, const PSDHeader& header);
+	bool parseLayer(QDataStream& s, const PSDHeader& header);
 	QString getLayerString(QDataStream & s);
 	void putDuotone(uchar *ptr, uchar cbyte);
